% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_columns.R
\name{infer_color_chem}
\alias{infer_color_chem}
\title{Infer colour chemistry from sequencer name}
\usage{
infer_color_chem(x)
}
\arguments{
\item{x}{character. The sequencer name as it appears in Pandora.}
}
\value{
integer
}
\description{
Infer colour chemistry from sequencer name
}
