% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_columns.R
\name{infer_library_specs}
\alias{infer_library_specs}
\title{Infer strandedness and udg_treatment from protocol number}
\usage{
infer_library_specs(x, pandora_library_protocol_info)
}
\arguments{
\item{x}{character. The libbary protocol as it appears in Pandora.}

\item{pandora_library_protocol_info}{Tibble with pandora protocol info}
}
\value{
character vector
}
\description{
Infer strandedness and udg_treatment from protocol number
}
