% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_library_protocol_info.R
\name{load_library_protocol_info}
\alias{load_library_protocol_info}
\title{Collect UDG and strandedness information for Pandora library preparation protocols}
\usage{
load_library_protocol_info(con)
}
\arguments{
\item{con}{A pandora database connection object.}
}
\value{
A tibble containg the protocol_name, udg, and strandedness of all library protocols in pandora.
}
\description{
Collect UDG and strandedness information for Pandora library preparation protocols
}
