% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_library_protocol_info.R
\name{validate_protocol_info}
\alias{validate_protocol_info}
\title{Validate pandora library info}
\usage{
validate_protocol_info(x)
}
\arguments{
\item{x}{tibble.}
}
\value{
tibble
}
\description{
Validate pandora library info
}
